Define the LineSegment data type
This completely removes the dependency on the methods
'free_ntuple_list', 'new_ntuple_list', 'enlarge_ntuple_list',
'add_7tuple', and the data structures 'ntuple_list_s' and 'ntuple_list'
from the original LSD source code. Instead, the output from the
algorithm in this package will be a slice of 'LineSegment' structs.

This new refactoring avoids extraneous (and confusion-causing)
generality in the original code.